backend/requirements.txt
anyio==3.6.2
attrs==23.1.0
black==23.3.0
bleach==6.0.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-log==0.4.0
colorama==0.4.6
cryptography==40.0.2
dataclasses==0.6
deprecation==2.1.0
docutils==0.20.1
dotty-dict==1.3.1
exceptiongroup==1.1.1
Flask==2.3.2
gitdb==4.0.10
GitPython==3.1.31
gotrue==0.2.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.12.3
httpx==0.16.1
idna==2.10
importlib-metadata==6.6.0
iniconfig==2.0.0
invoke==1.7.3
itsdangerous==2.1.2
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.2
keyring==23.13.1
MarkupSafe==2.1.2
more-itertools==9.1.0
mypy-extensions==1.0.0
numpy==1.24.3
packaging==23.1
pathspec==0.11.1
pkginfo==1.9.6
platformdirs==3.5.1
pluggy==1.0.0
postgrest==0.10.6
postgrest-py==0.4.0
py==1.11.0
pycparser==2.21
pydantic==1.10.8
Pygments==2.15.1
pytest==6.2.5
pytest-flask==1.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gitlab==3.14.0
python-semantic-release==7.33.2
readme-renderer==37.3
realtime==1.0.0
realtime-py==0.1.3
requests==2.25.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
SecretStorage==3.3.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starkbank==2.20.0
starkbank-ecdsa==2.2.0
starkcore==0.1.1
storage3==0.5.2
StrEnum==0.4.10
supabase==1.0.3
supabase-py==0.0.2
supafunc==0.2.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.8
tqdm==4.65.0
twine==3.8.0
typing_extensions==4.6.2
urllib3==1.26.16
webencodings==0.5.1
websockets==9.1
Werkzeug==2.3.4
zipp==3.15.0
